JVMCBC-357: Add raw N1QL querying msg that returns JSON
Motivation
----------
In some use cases, notably when developing a REST backend that could
just build N1QL queries and pass the JSON response back to a javascript
frontend, being able to perform a N1QL query and not interpret the
response but rather return is raw (JSON ByteBuf) can make sense.

Modifications:
-------------
Implemented a dedicated message for this that will just aggregate the
response body in the handler and return it. The HTTP status code and
message are also be returned. Note the new message extends the
GenericQueryRequest one.

Result
------
Users can directly query the N1QL service and get the uninterpreted
JSON response by using the RawQueryRequest message.

Change-Id: I9f96bb67da8132fa28b29e4bcfb59343f8545f32
Reviewed-on: http://review.couchbase.org/66937
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>