JVMCBC-360: View reduce can leak 1 byte of info()
Motivation:
-----------
When a view is queried with reduce = true, the `total_rows` section can
be omitted by the server. The way the handler reacts to that is that it
stops using the ByteBuf it prepared for info() and instead emit an empty
buffer (`Unpooled.EMPTY_BUFFER`).

The thing is that the ByteBuf that was prepared is never released,
leading to a 1 byte leak (the opening bracket from the response).

Modification:
-------------
ViewHandler now correctly releases the info ByteBuf if it is not emitted
to the upper layers.

Result:
-------
The leak has been eliminated.

Change-Id: I2529a36001540a045ebde98eebb283bc1a04f4db
Reviewed-on: http://review.couchbase.org/67021
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>