JVMCBC-365: Fail with better exception on InvalidKeyException.
Motivation
----------
On JDKs like IBM JDK or older OpenJDK/Hotspots (small 7, all 6)
the underlying code does not accept an empty string to initialize
an HMAC, but this is required for our new SCRAM authentication
approach.

This only happens on empty bucket passwords as well.

Modifications
-------------
The code now explicitly fails with an exception telling the user
what needs to be done to troubleshoot.

In addition, disabling SCRAM with a system property is also possible
but it is not mentioned since it is considered an insecure workaround
and should only be done if not possible otherwise.

Result
------
If a condition like this happens the user will know what is going
on.

Change-Id: Id0d36eeb516012615cad73851c167b2579fbe1f2
Reviewed-on: http://review.couchbase.org/68088
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>