JCBC-1007: Fix queries not load balancing correctly when in MDS.
Motivation
----------
The round robin mechanism for locating the proper nodes on Query, FTS and Views
correctly used an "offset" when iterating through the node list, but the list
also includes nodes with different services. In a setup like this:

+------------+------------+---------------+---------------+
| Node A: KV | Node B: KV | Node C: Query | Node D: Query |
+------------+------------+---------------+---------------+

Round robin would skip forward on the nodes where query is not enabled, leading
to queries being sent to: C, C, C, D, C, C, C, D, since C would handle offsets
for A and B respectively. In non-MDS this is not an issue since query nodes
are available on all machines. In other constellations this is not an issue
as well, but the ordering can't be controlled by the user and as a result
is non-deterministic.

Query service is illustrated but the same issue is found in FTS and Views.

Modifications
-------------
Unit tests have been added for this situation as regression tests. After considering
ideas to advance the offset in various ways I determined the safest way to get
to a good approach is to filter the nodes list upfront to only include nodes
for this particular service. Then in all cases we can be sure only nodes for
that specific service are touched in the first place.

Result
------
Since now the nodes are filtered by the service type, other nodes with different
services do not skew the round-robin algorithm and as a result n1ql queries
(as well as view and fts ones) are distributed more evenly over the cluster.

Change-Id: I4e178b6fdb7508b95e198d703b43a7f30bd08213
Reviewed-on: http://review.couchbase.org/68148
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>