JVMCBC-376: Add support for legacy memcached node hashing.
Motivation
----------
The node hashing for ketama in the 2.x SDK series is different
than in the 1.x series. Because of this, using the 1.x and 2.x
series togther for memcached buckets is incompatible.

To make both work together this changeset makes it possible
to plug in a "legacy" node hashing mechanism that is
compatible with the 1.x series.

Modifications
-------------
This changeset refactors out the node hashing from the memcached
config into an interface and 2 implementations, sticking with the
previous one for backwards compatibility.

The new method is exposed on the environment and can be optionally
enabled. To make this work the config needs to have the environment
in scope so the code is modified to inject it when decoding.

Result
------
It is possible to optionally enable a legacy decoding mechanism
that makes it compatible with java 1.x SDK as well as all the other
drivers that use the same scheme.

Change-Id: I0e13e1aa12cdd3412d30b6910f914d1f51dce5e3
Reviewed-on: http://review.couchbase.org/70294
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>