Bump Dependencies for 1.4.0.
Motivation
----------
As with every minor release it makes sense to refresh our
dependencies, including RxJava and netty.

Note that this Jackson is also bumped from 2.6.x to 2.7.x. According
to their documentation 2.7 is still compatible (runnable) with java
1.6, but can only be compiled with 1.7. The main reason for the bump
is that 2.6 didn't have releases in a while and since JSON is such
an integral part of the SDK it makes sense to keep it safe and
speedy.

  - RxJava to 1.2.3
  - Netty to 4.0.42
  - Disruptor to 3.3.6
  - Jackson to 2.7.8

Change-Id: I621c9b3da1f86914135add955b004e1cd5f416f7
Reviewed-on: http://review.couchbase.org/70864
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>