Revision of the Exceptions constructors with no params
Motivation
----------
All the java client exceptions having a a method with no parameters dont make a super call.
The oracle compiler in compilation time add this call but all the exceptions have as father
the CouchbaseException and his no params constructor nor have the super call.

This is a problem because the stack trace is never initialized.

Modifications
-------------
Add super() to all the constructors.

Result
------
The Exception is correctly initialized.

Change-Id: I6b2b6fbc76b17d95e780948a14c5a1ff46b855b6
Reviewed-on: http://review.couchbase.org/71187
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>