JVMCBC-373: Move onError calls onto scheduler.
Motivation
----------
Like with onNext calls, it is important that (if not configured
to do so explicitly) onError callbacks are also executed on the
scheduler (and not run directly on the threads where called).

Modifications
-------------
The code adds a helper class which contains all the conditional
logic to make sure the observables is failed on the proper scheduler
and the rest of the code is adapted (where it makes sense) to
call this helper method instead of failing the observable dircectly.

Result
------
The net result is that error callbacks are now also moved to the
proper scheduler pools, like onNext calls.

Change-Id: I4a1e167adb57bb92e8477fba6b2f88356f284139
Reviewed-on: http://review.couchbase.org/71334
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>