JVMCBC-381: No keepalive for non-pipelined endpoints when op is in-flight
Motivation
----------
On non-pipelined endpoints it is important to not send keepalives when an
operation is in-flight, since sending the keepalive has the same effect
as pipelining another op over that socket, leading to prematurely termination
of the previous query.

Modifications
-------------
The AGH has been modified to only initiate the keepalive sending process if
either pipelining is enabled (then we always send it) or otherwise no operation
is either queued in the sentRequestQueue or currently being decoded. This
is similar to the check previously added on regular ops under non-pipelining.

Result
------
Pipelining does not implicitly occur on non-pipeline enabled endpoints with
keepalives.

Change-Id: I2b660895e6548464c5e28f09195b4083838eaaa0
Reviewed-on: http://review.couchbase.org/71800
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-on: http://review.couchbase.org/71963