JVMCBC-382: Extract interface needed for BucketConfigParser
Motivation
----------
Right now it is necessary to implement CoreEnvironment to parse
the configuration with BucketConfigParser.parse, which is
not really necessary, and ClientEnvironment in JDCP does not
implement full CoreEnvironment.

Modification
------------
Extract minimal interface, needed for parser and inherit in it
CoreEnvironment.

Result
------
JDCP could inherit ConfigParserEnvironment too in its ClientEnvironment.

Change-Id: I52d254270b7825dbe1c01c0305f5f40ce769978f
Reviewed-on: http://review.couchbase.org/71803
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>