JVMCBC-383: Redistribute BucketConfigRequest if hostname not set.
Motivation
----------
If the BucketConfigRequest does not have a hostname explicitly
attached (which is needed during bootstrap), the code should
apply the usual round-robin approach to make sure that even
if a node is down the config message succeeds eventually.

Modifications
-------------
The code is modified so that a BucketConfigRequest without
a hostname just goes through the same round-robin approach
as other config messages. Previously, always the same node
would've been used so up-the-stack retry mechanisms were
fruitless since the same node would be tried over and over
again for the BucketConfigRequest.

Test has been added to assert this.

Result
------
The BCR can be properly dispatched, even when a node is
down.

Change-Id: I99cc53ec3e561668435af9e7e2ff908a7be0c249
Reviewed-on: http://review.couchbase.org/71898
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-on: http://review.couchbase.org/71964