JVMCBC-385: Allow to fall back to old services.
Motivation
----------
Since the new services are implemented in a bugfix release
there should be an escape valve that allows to force and go
back to the old services in case a bug has slipped through.

Modifications
-------------
This changeset simply introduces the old services and modifies
the factory to use them if the com.couchbase.forceOldServices
system property is set to true.

This should not be documented and only used to advice a user
to circumvent a bug if it slipped through testing.

Result
------
We now have an escape valve if a user is running into troubles
with the new services implementation.

Change-Id: I8ee68093f1d4e7887a40ef8c4c3550d2eac2eb02
Reviewed-on: http://review.couchbase.org/73193
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>