JVMCBC-395: support ephemeral buckets in config.
Motivation
----------
Since Couchbase Server 5.0 introduces ephemeral buckets
the SDK needs to support it. Specifically, this has impact
on two pieces.

  - PersistTo can't work by definition
  - Views are inherently disk-based so they also don't work

Modifications
-------------
The server removes parts from the bucket config when ephemeral
bucket mode is selected, namely views altogether. Since prev.
views always implied kv service, the code now adapts to the
fact that views might be missing.

For PersistTo we check if the config is ephemeral (for this a
new getter has been added to the couchbase bucket config) and
if so we short-circuit the observe chain.

This changeset also adds parsing for the bucket capabilities,
but doesn't use it if it is null for backwards compatibility
reasons.

Result
------
Support for ephemeral buckets.

Change-Id: I1da500838239959b41c615e0f57d14ef41963436
Reviewed-on: http://review.couchbase.org/74315
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>