JVMCBC-393: Add support for the extended KV error map.
Motivation
----------
To be more forwards compatible with future error codes the
server adds a new capability called "the error map", which
is outlined in the internal server documentation.

As a first step, being disabled by default, the client should
be able to load and store this error map and still log
all unknown errors. In future changes the new capabilities
will be enhanced to actually make use of specific behavior.

Modifications
-------------
This changeset adds a new KV handler during bootstrap which,
if enabled, negotiates the XERROR map via HELLO and if set
then adds a new handler that fetches the error map itself.

If loaded properly, the ResponseStatusConverter is fed with
the map but it is only stored if the new map is actually newer
than the currently known one.

For now, unknown errors are looked up in the map and printed
if enabled, but the user won't see a change in behavior
for now.

Result
------
Basic support for the new KV error map feature.

Change-Id: Iaefa0787c19d66c8439b0db2ab4ffb746171e9ad
Reviewed-on: http://review.couchbase.org/74847
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>