JVMCBC-401: propagate error on endpoint encode.
Motivation
----------
If an error happens on encode in the IO pipeline
it is swallowed and turns into a timeout on the
caller side - but it instead should be propagated
back through the request so the user knows whats
going on.

Modifications
-------------
This changeset catches an error during the encoding
phase and if this happens, it fails the request
(which in turn will propagate it back to the user)
instead of failing it silently and never completing.

A test has been added to catch regressions in the
future.

Result
------
Correctly catch and propagate errors on the IO
encode path.

Change-Id: I85599be36e13704693006394b42c0f113375b784
Reviewed-on: http://review.couchbase.org/75428
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Michael Nitschinger <michael@nitschinger.at>