JVMCBC-409: Resolve addresses when parsing ConnectionString
Motivation
==========
Otherwise it might be get resolved internally when it is too late.
Or netty will raise error if it will see InetSocketAddress, which
is not resolved.

Also before JDK 7+ it is not possible to get original value of the
hostname, and using getHostName() might trigger reverse-lookup if
it was initialized by raw IP address.

Modification
============
Use constructor when adding new hosts in the parser, which will
result in resolved InetSocketAddress.

Result
======
Behaviour will be more determined, and user will see any name errors
earlier.

Change-Id: I4f669d9d76218f3909e11a6430da29bd4230d1a2
Reviewed-on: http://review.couchbase.org/76061
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>