JVMCBC-410: Allow to configure the config polling interval.
Motivation
----------
At the moment the config poll interval is fixed to 10 seconds,
but it should be configurable.

Modifications
-------------
This changeset exposes the setting on the environment and makes
use of it in the CarrierRefresher. Note that when set to 0, it
is considered to be disabled.

Also, the interval cannot be configured lower than 2500
milliseconds to prevent the user from polling at a very high
rate all the time and potentiall overloading the system.

Result
------
Configurable poll interval, with the option to disable it if
needed.

Change-Id: Ib0180082a83422043e9e9aa6f4d07e4cd5c8033a
Reviewed-on: http://review.couchbase.org/76064
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>