JVMCBC-394: Fire channelActive in SelectBucketHandler when disabled.
Motivation
----------
As part of the bucket selection process a new handler was introduced,
but it doesn't fire the channel active event if the feature is
disabled. Since the upper handlers rely on this event to be bubbled
up the code doesn't work properly anymore.

Modifications
-------------
Bubble up the channelActive event if the feature is disabled.

Result
------
Functionality restored.

Change-Id: I922a87987589f0b0d719e1a2e33155abcd0f714f
Reviewed-on: http://review.couchbase.org/76183
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>