JVMCBC-394 Apply username to newConfig in ClusterConfig bucketconfigs
Motivation
----------
To support RBAC username of the user authorized for bucket access is
included in the bucket configuration. The username in the bucket
configuration is not applied to new configs received.

Changes
-------
Since the username is not set on new cluster configuration received on
cluster change, this could lead to auth failures in the key value endpoint
negotiation as the username is pulled from the configuration. So the
change apply the username from the old config to new configurations
received, as the cluster configuration will not by itself contain that
information.

Results
-------
No auth failures on cluster change.

Change-Id: I4a0249f91b369ff640a9a960f250a892e9003510
Reviewed-on: http://review.couchbase.org/76213
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>