JVMCBC-413: Allow to run HELLO after auth.
Motiviation
-----------
To support EOL server versions it should be possible
to optionally run HELLO after auth (to prevent the
server disconnecting the client over and over again).

Modifications
-------------
This change adds a system property called
com.couchbase.helloAfterAuth, which defaults to
false. If set to true, the pipeline is switched
around to first run the authentication phase
and later hello.

The downside to this is that extended errors on auth
are not available, but this is not a concern with older
servers anyways.

Result
------
Backwards compatibility with older, EOL server versions.

Change-Id: I1a723ed3ae207447c62a58eac530cf1660cf556d
Reviewed-on: http://review.couchbase.org/76559
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>