JVMCBC-415: Enforce CarrierRefresher poll floor.
Motivation
----------
As part of "fast failover" capabilities it has been decided that a
global polling floor of 10ms should be enforced in order to make
sure that the server is not flooded with config requests.

Modifications
-------------
Because all polling is directed through the CarrierRefresher in the
core-io layer, a volatile timestamp has been introduced which is
now updated and checked every time a poll (either through refresh
or tainted pollign) needs to be initiated.

A test has been added which polls quickly and asserts that only
every 10ms or more the actual polling call is directed against
the servers.

Result
------
A floor polling of 10ms is introduced, making sure that the
servers are not flooded with polling requests.

Change-Id: I71e2bf8ab5389d4a0baba811c5a8cff4b1ae5fae
Reviewed-on: http://review.couchbase.org/76845
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>