JVMCBC-421: Close LatencyStats on removal.
Motivation
----------
LatencyStats is holding a weak reference to the pause detector
and it makes sense to clean them up to save the GC some work
and also avoid bloating if the GC is not cleaning them up
right away.

Modifications
-------------
Once the stat is removed, call close on it to make sure the
weak reference internally is cleaned up properly.

Result
------
LatencyStats are properly cleaned up. Hat tip to @mp911de for
letting me know of this.

Change-Id: Ic7e8640831cbdb06c2e7dc434567654382ba4208
Reviewed-on: http://review.couchbase.org/77237
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>