JVMCBC-423 Dont send timed out operations over network
Motivation
----------
It is not very ideal to send timed out operations on the application to
the server.

Changes
-------
Allow to store a reference to the subscriber in the CouchbaseRequest and
add check to see if the subscriber has unsubscribed from the subscriptions list.
If the subscriber has unsubscribed, the request handler will not further
process the request.

Results
-------
Timed out requests will not be sent to the server. Tests are covered
in java-client.

Change-Id: Ied419146c6a22f6041db7f7e29e40924f68b7585
Reviewed-on: http://review.couchbase.org/77515
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>