JVMCBC-422: Add utility class for inet address management.
Motivation
----------
In prepraration for bigger changes to the code base around
inet addr management, the code needs a helper/wrapper class
around InetAddress to be more intelligent about reverse
DNS lookups and in the future other functionality.

Modifications
-------------
As a result, this changeset introduces a NetworkAddress class
which acts as a wrapper around InetAddress and has more knowledge
about the environment as well as how it has been initialized.

Based on this it provides accessors to get to the hostname,
ip address or a combination in a best-effort manner. It is intended
to be used troughout the codebase.

In addition utility methods are imported through InetAddresses, copied
from Google Guava with proper attribution in the header.

Result
------
We now have a wrapper class around InetAddress which is more intelligent
and allows us to have proper orchestration around it.

Change-Id: Ic2cfbe3b3dbcb2aa2519a20aa11b3a0ac275edda
Reviewed-on: http://review.couchbase.org/78182
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>