JVMCBC-422: Adapt codebase to use the NetworkAddress.
Motivation
----------
The previous change did introduce the NetworkAddress wrapper,
this change modifies the codebase to make use of it.

Modifications
-------------
Code and tests are adapted, various interfaces changed to expose
the new signature. Note that for backwards compat on external
interfaces, the event bus message getters have not been changed.

Result
------
The code now makes use of the address wrapper and allows more
control over the reverse DNS behavior.

Change-Id: I361fd52abd3722956fbbf53868fc9aa6d70d40a4
Reviewed-on: http://review.couchbase.org/78236
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>